% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{accrint}
\alias{accrint}
\title{Accrued Interest for Periodic Coupon Security}
\usage{
accrint(issue, first_interest, settlement, rate, par, frequency, basis)
}
\description{
Accrued Interest for Periodic Coupon Security
}
