% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{amordegrc}
\alias{amordegrc}
\title{French Degressive Depreciation}
\usage{
amordegrc(cost, date_purchased, first_period, salvage, period, rate, basis)
}
\description{
French Degressive Depreciation
}
