% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{amorlinc}
\alias{amorlinc}
\title{French Straight-Line Depreciation}
\usage{
amorlinc(cost, date_purchased, first_period, salvage, period, rate, basis)
}
\description{
French Straight-Line Depreciation
}
