% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_functions-package.R
\docType{package}
\name{FinancialFunctions-package}
\alias{FinancialFunctions-package}
\alias{FinancialFunctions}
\title{FinancialFunctions: Excel-Compatible Financial Functions Powered by Rust}
\description{
Provides 55 Excel-compatible financial functions implemented in Rust
for high performance. All functions handle invalid inputs gracefully
by returning NA rather than throwing errors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Tor-Storli/Financial_Functions}
  \item Report bugs at \url{https://github.com/Tor-Storli/Financial_Functions/issues}
}

}
\author{
\strong{Maintainer}: Tor Storli \email{your.email@example.com}

Authors:
\itemize{
  \item Tor Storli \email{your.email@example.com}
}

}
