% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{intrate}
\alias{intrate}
\title{Interest Rate for Fully Invested Security}
\usage{
intrate(settlement, maturity, investment, redemption, basis)
}
\description{
Interest Rate for Fully Invested Security
}
