% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{oddfprice}
\alias{oddfprice}
\title{Price with Odd First Period}
\usage{
oddfprice(
  settlement,
  maturity,
  issue,
  first_coupon,
  rate,
  yld,
  redemption,
  frequency,
  basis
)
}
\description{
Price with Odd First Period
}
