% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{oddfyield}
\alias{oddfyield}
\title{Yield with Odd First Period}
\usage{
oddfyield(
  settlement,
  maturity,
  issue,
  first_coupon,
  rate,
  pr,
  redemption,
  frequency,
  basis
)
}
\description{
Yield with Odd First Period
}
