% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{oddlprice}
\alias{oddlprice}
\title{Price with Odd Last Period}
\usage{
oddlprice(
  settlement,
  maturity,
  last_interest,
  rate,
  yld,
  redemption,
  frequency,
  basis
)
}
\description{
Price with Odd Last Period
}
