% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{oddlyield}
\alias{oddlyield}
\title{Yield with Odd Last Period}
\usage{
oddlyield(
  settlement,
  maturity,
  last_interest,
  rate,
  pr,
  redemption,
  frequency,
  basis
)
}
\description{
Yield with Odd Last Period
}
