% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pricemat}
\alias{pricemat}
\title{Price of Security Paying Interest at Maturity}
\usage{
pricemat(settlement, maturity, issue, rate, yld, basis)
}
\description{
Price of Security Paying Interest at Maturity
}
