% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{vdb}
\alias{vdb}
\title{Variable Declining Balance Depreciation}
\usage{
vdb(cost, salvage, life, start_period, end_period, factor, no_switch)
}
\description{
Variable Declining Balance Depreciation
}
