% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{yieldmat}
\alias{yieldmat}
\title{Yield of Security Paying Interest at Maturity}
\usage{
yieldmat(settlement, maturity, issue, rate, pr, basis)
}
\description{
Yield of Security Paying Interest at Maturity
}
